refactor: update import paths for auth and prisma modules
- Changed import paths for `authOptions` and `prisma` from `@/app/lib` to `@/lib` across multiple API routes and components for consistency and clarity.
- Removed unused `auth.ts` and `prisma.ts` files from the `lib` directory to streamline the project structure.
