+++ app/api/config/scoring/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server"
 import { getServerSession } from "next-auth"
-import { authOptions } from "@/app/lib/auth"
-    import prisma from "@/app/lib/prisma"
+import { authOptions } from "@/lib/auth"
+    import prisma from "@/lib/prisma"
 
 export async function GET() {
   try {
