+++ app/api/debug/server-info/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { prisma } from "@/app/lib/prisma";
+import { prisma } from "@/lib/prisma";
 
 export async function GET() {
   let databaseStatus = { connected: false, type: "PostgreSQL" };
