+++ app/api/face-detection/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/lib/auth';
-import prisma from '@/app/lib/prisma';
+import { authOptions } from '@/lib/auth';
+import prisma from '@/lib/prisma';
 
 // Face detection response type
 interface HandRaising {
