+++ app/api/user/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server"
-import { prisma } from "@/app/lib/prisma"
+import { prisma } from "@/lib/prisma"
 import { getServerSession } from "next-auth"
-import { authOptions } from "@/app/lib/auth"
+import { authOptions } from "@/lib/auth"
 
 export async function GET() {
   try {
